기능: DataFieldCollectable 로 추상 데코레이터의 주요 기능을 모듈로 추출
